+++ protocols/gossipsub/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.46.1 - unreleased
+- Implement publish and forward message dropping.
 
 - Implement backpressure by diferentiating between priority and non priority messages.
   Drop `Publish` and `Forward` messages when the queue becomes full.